move classes and rename fields in response to PR review
also removed some Bundle updates that are no longer necessary
