+++ src/main/java/com/conveyal/analysis/controllers/UserActivityController.java
@@ -2,12 +2,8 @@ package com.conveyal.analysis.controllers;
 
 import com.conveyal.analysis.UserPermissions;
 import com.conveyal.analysis.components.TaskScheduler;
-import com.conveyal.analysis.models.BaseModel;
-import com.conveyal.analysis.models.Bundle;
-import com.conveyal.analysis.models.Model;
-import com.conveyal.analysis.models.OpportunityDataset;
-import com.conveyal.analysis.models.RegionalAnalysis;
 import com.conveyal.r5.analyst.progress.Task;
+import com.conveyal.r5.analyst.progress.WorkProduct;
 import spark.Request;
 import spark.Response;
 import spark.Service;
@@ -71,46 +67,9 @@ public class UserActivityController implements HttpController {
         public String detail;
         public Task.State state;
         public int percentComplete;
-        public Long timeBegan;
-        public Long timeCompleted;
+        public int secondsActive;
+        public int secondsComplete;
         public WorkProduct workProduct;
-    }
-
-    /** API model providing a unique identifier for the final product of a single task in an activity response. */
-    // ID is unique within type so region is redundant information, but facilitates prefectch on API.
-
-    public static class WorkProduct {
-        public WorkProductType type;
-        public String id;
-        public String region;
-
-        public WorkProduct (WorkProductType type, String id, String region) {
-            this.type = type;
-            this.id = id;
-            this.region = region;
-        }
-
-        public static WorkProduct forModel (Model model) {
-            // FIXME Not all Models have a regionId. Rather than pass that in as a String, refine the programming API.
-            WorkProduct product = new WorkProduct(WorkProductType.forModel(model), model._id, null);
-            return product;
-        }
-    }
-
-    /**
-     * There is some implicit and unenforced correspondence between these values and those in FileCategory, as well
-     * as the tables in Mongo. We should probably clearly state and enforce this parallelism. No background work is
-     * done creating regions, projects, or modifications so they don't need to be represented here.
-     */
-    public static enum WorkProductType {
-        BUNDLE, REGIONAL_ANALYSIS, AGGREGATION_AREA, OPPORTUNITY_DATASET;
-        // Currently we have two base classes for db objects so may need to use Object instead of BaseModel parameter
-        public static WorkProductType forModel (Model model) {
-            if (model instanceof Bundle) return BUNDLE;
-            if (model instanceof OpportunityDataset) return OPPORTUNITY_DATASET;
-            if (model instanceof RegionalAnalysis) return REGIONAL_ANALYSIS;
-            throw new IllegalArgumentException("Unrecognized work product type.");
-        }
     }
 
 }
