+++ src/main/java/com/conveyal/r5/analyst/progress/ProgressInputStream.java
@@ -1,6 +1,5 @@
-package com.conveyal.analysis.util;
+package com.conveyal.r5.analyst.progress;
 
-import com.conveyal.r5.analyst.progress.ProgressListener;
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.io.input.ProxyInputStream;
 
@@ -11,9 +10,10 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static org.apache.commons.io.IOUtils.EOF;
 
 /**
- * This will report the total number of bytes that have passed through the stream, which can exceed 100% of the file
- * size if the caller uses mark and reset. Based on CountingInputStream. The progressListener should be pre-configured
- * with the total number of bytes expected and a detail message using ProgressListener::beginTask.
+ * This will report progress as the total number of bytes that have passed through the stream, like CountingInputStream.
+ * This can exceed 100% of the file size if the caller uses mark and reset. The progressListener should be
+ * pre-configured with the total number of bytes expected and a detail message using ProgressListener::beginTask.
+ * The static method forFileItem() demonstrates usage when reading from a file of known length.
  */
 public class ProgressInputStream extends ProxyInputStream {
 
@@ -42,7 +42,7 @@ public class ProgressInputStream extends ProxyInputStream {
      * Given an uploaded file, report progress on reading it.
      * Incrementing the progress seems to introduce some inefficiency when performing unbuffered small reads, such as
      * calls to InputStream.read() which are used by DataInputStream to read numbers.
-     * TODO wrap in buffered input stream to reduce small read calls? Or make tune progress reporting?
+     * TODO wrap in buffered input stream to reduce small read calls, or tune to only report once per percentage?
      */
     public static ProgressInputStream forFileItem (FileItem fileItem, ProgressListener progressListener) {
         try {
